fix code preview
