+++ app/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+from . import errors
+from . import routes
+
+# create application instance
+app = Flask(__name__)
