+++ requirements.txt
@@ -3,3 +3,5 @@ celery>=4.3
 kombu>=4.5.0
 pytz>=2015.4
 pymongo>=3.7.2
+cerberus==1.2
+pytest
